refactor(schemas): migrate Phase 36 endpoint schemas — 6 files, 12 classes (#6042)
Move local Pydantic class definitions from endpoint files into matching
domain schema modules.

- integration_version_control.py: ConnectionTestRequest, ProviderInfo →
  schemas_workflows.py
- kb_librarian.py: KBQuery, KBQueryResponse → schemas_knowledge.py
- knowledge_audit.py: AuditEvent, ComplianceReportRequest →
  schemas_knowledge.py (AuditEventType imported from knowledge.audit_log)
- onboarding.py: ApplyPresetRequest, OnboardingStatus → schemas_system.py
- orchestration.py: WorkflowRequest, AgentRecommendationRequest →
  schemas_workflows.py
- analytics_debt.py: DebtCalculationRequest, DebtSummary → schemas_analytics.py
  (DebtCategory/DebtSeverity enums + DebtItem dataclass kept in endpoint
  because they are tightly coupled with calculation helpers)

Endpoint files now import from the schemas modules; no behavior change.
