+++ autobot-backend/api/integration_version_control.py
@@ -8,19 +8,19 @@ import logging
 from typing import Any, Dict, List, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query
-from pydantic import BaseModel, Field
 
-from auth_middleware import check_admin_permission
-from integrations.base import IntegrationConfig, IntegrationHealth
-from integrations.version_control_integration import (
-    BitbucketIntegration,
-    GitLabIntegration,
-)
 from api.schemas_code import (
     VCSBranchesResponse,
     VCSCommitInfoResponse,
     VCSPullRequestsResponse,
     VCSRepositoriesResponse,
+)
+from api.schemas_workflows import ConnectionTestRequest, ProviderInfo
+from auth_middleware import check_admin_permission
+from integrations.base import IntegrationConfig, IntegrationHealth
+from integrations.version_control_integration import (
+    BitbucketIntegration,
+    GitLabIntegration,
 )
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 
@@ -29,30 +29,6 @@ router = APIRouter(
     tags=["integrations-version-control"],
     dependencies=[Depends(check_admin_permission)],
 )
-
-
-class ConnectionTestRequest(BaseModel):
-    """Request model for testing VCS connection."""
-
-    provider: str = Field(..., description="VCS provider (gitlab, bitbucket)")
-    api_key: str = Field(..., description="API key or access token")
-    settings: Dict[str, Any] = Field(
-        default_factory=dict, description="Provider-specific settings"
-    )
-
-
-class ProviderInfo(BaseModel):
-    """Information about a VCS provider."""
-
-    id: str = Field(..., description="Provider identifier")
-    name: str = Field(..., description="Provider display name")
-    description: str = Field(..., description="Provider description")
-    required_settings: List[str] = Field(
-        default_factory=list, description="Required configuration settings"
-    )
-    optional_settings: List[str] = Field(
-        default_factory=list, description="Optional configuration settings"
-    )
 
 
 SUPPORTED_PROVIDERS = {
